move change role dialog from members to members table
